+++ addon/components/select-2.js
@@ -32,6 +32,7 @@ var Select2Component = Ember.Component.extend({
   // Bindings that may be overwritten in the template
   inputSize: "input-md",
   cssClass: null,
+  optionIdPath: "id",
   optionValuePath: null,
   optionLabelPath: 'text',
   optionHeadlinePath: 'text',
@@ -59,6 +60,7 @@ var Select2Component = Ember.Component.extend({
   didInsertElement: function() {
     var self = this,
         options = {},
+        optionIdPath = this.get('optionIdPath'),
         optionLabelPath = this.get('optionLabelPath'),
         optionHeadlinePath = this.get('optionHeadlinePath'),
         optionDescriptionPath = this.get('optionDescriptionPath'),
@@ -79,7 +81,7 @@ var Select2Component = Ember.Component.extend({
 
     // override select2's default id fetching behavior
     options.id = (function(e) {
-      return (e === undefined) ? null : get(e, 'id');
+      return (e === undefined) ? null : get(e, optionIdPath);
     });
 
     // allowClear is only allowed with placeholder
@@ -103,7 +105,7 @@ var Select2Component = Ember.Component.extend({
       }
 
       var output,
-          id = get(item, "id"),
+          id = get(item, optionIdPath),
           text = get(item, optionLabelPath),
           headline = get(item, optionHeadlinePath),
           description = get(item, optionDescriptionPath);